Revert "Fix messagebus warnings"
This reverts commit 52c5caae432586d101f9046d32665695a8c4de0b.

+++ include/etl/shared_message.h
@@ -174,7 +174,7 @@ namespace etl
     //*************************************************************************
     ETL_NODISCARD uint32_t get_reference_count() const
     {
-      return static_cast<uint32_t>(p_rcmessage->get_reference_counter().get_reference_count());
+      return p_rcmessage->get_reference_counter().get_reference_count();
     }
 
     //*************************************************************************
